Add backup feature and .gitignore file
A new backup feature has been added to the project. The backup.go file contains the implementation for this feature. It also introduces new mandatory flags for the backup command. The newly introduced backup and start.go files have been attached to the root command in their respective init functions. In addition to this, .gitignore file has been introduced to the project to ignore unnecessary files and folders. It mainly ignores the binary files and vendor dependencies. The version.go file provides a version command to the cli tool. The required dependencies have been added to go.mod and their respective checksums to go.sum.