Add destination file fallback and listing functionality
The change introduces a fallback mechanism for the destination file path, in case it's not set in environment variables, a default location is used (".klusoga-backup/destinations.yaml" in the user's home directory). We also ensure that the file and its directories exist, and if not, create them.

In addition to this, a new method "ListDestinations" was added to return all destinations specified in the file. This function gets used in a new command "destinations", listing all defined destinations to the console.

The `Destination` interface now includes a `String` method, too, which provides a convenient way of printing destination information.

This is helpful in situations where the destinations are dynamically managed and gives more transparency to the backup agent's configuration.