Change bucket parameter for S3 file upload
Modified bucket parameter in the PutObjectInput method from static 's.params.Bucket' to dynamic 'prefix[0]' in aws.go file. This change ensures the bucket name is not hard-coded and it provides flexibility for automatic prefix generation based on the given bucket name which brings in increased adaptability and code efficiency.