Update entrypoint and copy command in Dockerfile
The commit updates the ENTRYPOINT and COPY directives in the Dockerfile. The changes were made to reflect the name change of our primary application from 'klusoga-backup-agent' to 'klusoga-backup'. This ensures that our Docker containers are initialized correctly using the updated application name. By correctly referencing to the new application name, the newly built Docker containers will work as expected.